[Author: thakis]
Unset focus when hiding a per-tab sheet.

This is necessary so that when switching to a new tab using the
cmd-alt-right keyboard shortcut in chromium, the sheet in the previous
tab doesn't keep keyboard focus (which makes breaks cmd-l etc until the
window is clicked -- annoying).

R=thomasvl
DELTA=5  (5 added, 0 deleted, 0 changed)