[Author: oster]
subtreeDescription is a debugging tool for displaying UIView hierarchies in the Xcode Console window in DEBUG builds.
As it says in the .h file:
pause in the debugger and type something like:
(gdb) po [[[UIApplication sharedApplication] keyWindow] subtreeDescription]
and a nicely pretty printed version of the view hierarchy will be printed in the debugger console window.

DELTA=179 (179 added, 0 deleted, 0 changed)
R=dmaclach,mikemorton,thomasvl