[Author: thomasvl]
Support the toolbar item's view in state logging.
Add table columns to ui localizer (tests will come in a following cl).
Add a toolbar item that doesn't localize to help catch bugs in the future like we had on first impl.
Add a tooltip to the toolbar item in the unittest for ui localizer.
Add a note about the limits to ui localizer and toolbars due to the NSToolbar api's limits.

R=dmaclach
DELTA=746  (90 added, 599 deleted, 57 changed)