[Author: dmaclach]
This may help clean up some code in chrome, and make doing animations on
the mac easier in general. Not really sure how to unit test effectively, but
I'll try and come up with something this evening. Hopefully you will get a 
chance to look at it, so Andy can use it for his current CL in flight
http://codereview.chromium.org/1118004.

R=thomasvl
DELTA=908  (908 added, 0 deleted, 0 changed)